requirements.txt
# Django Framework
Django==5.1.7

# Admin Interface
django-jazzmin==3.0.0

# Forms
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0

# Authentication
django-allauth==65.3.0

# API
djangorestframework==3.15.2

# Utils
django-cleanup==9.0.0
django-debug-toolbar==4.4.6

# Database
# SQLite is included with Python, no additional package needed

# ASGI/WSGI servers (optional, for production)
# gunicorn==23.0.0
# uvicorn==0.32.1

# Image handling
Pillow>=10.0.0

# Testing
pytest==8.3.3
pytest-django==4.9.0
